NME
The enemies always look in front of them to see if there is an obstacle. When they encounter the obstacle, they change direction based on a random probability. They have a timer before they can start chasing Pacman around the map. Since they are the ones with the trigger collision, they are the ones calling Pacman's function DeathEvents. If they were only having a collision sphere, they wouldn't be real ghosts roaming around.